refactor(rich-content): centralize runtime and fixtures
