+++ features/rich-content/fixtures/README.md
@@ -1,11 +1,26 @@
 # Fixture 约定
 
+## 目录分层
+
+- `inbox/`：新样本收件箱。可以直接放任意数量的 `.md`、`.html` 或 `.htm`，也可以按来源自行建子目录；不要求先改 manifest。
+- `cases/`：已经挑选、命名并纳入稳定回归的样本。
+- `manifest.json`：只登记 `cases/` 中需要精确断言的案例。
+
+日常采集时先把文件放进 `inbox/`，运行：
+
+```bash
+npm run analyze:rich-content:inbox
+```
+
+工具会递归扫描每个文件，并把文件里的非空内容块分别列出来。这个阶段只做结构统计，不会因为尚未登记期望值而失败。
+
 ## 新增真实内容
 
-1. 将原始正文 HTML 保存到本目录，推荐命名：`<类型>-<主要特征>-<序号>.html`。
-2. 在 `manifest.json` 中添加案例；同一文件有多个非空块时使用 `block` 指定从 0 开始的块序号。
-3. 填写 `traits`，并只为已经人工确认的结构添加 `expected` 精确计数。
-4. 运行 `npm run analyze:rich-content` 和 `npm run test:rich-content`。
+1. 将原始正文放入 `inbox/`，推荐命名：`<类型>-<主要特征>-<序号>.md`。
+2. 用 inbox 分析命令检查段落、图片、公式、视频等结构。
+3. 确认样本有长期价值后，将它移动到 `cases/`，并在 `manifest.json` 中添加案例；同一文件有多个非空块时使用 `block` 指定从 0 开始的块序号。
+4. 填写 `traits`，并只为已经人工确认的结构添加 `expected` 精确计数。
+5. 运行 `npm run analyze:rich-content` 和 `npm run test:rich-content`。
 
 样本应尽量保留知乎返回的真实标签和属性，以便覆盖懒加载图片、公式、视频、链接卡片和异常嵌套。允许替换与结构无关的正文文本，但不要手工“修好”原始 HTML。
 
@@ -18,4 +33,6 @@
 
 ## 现有种子样本
 
-仓库根目录的 `lala.md` 是第一批历史样本，目前包含两个非空 HTML 块。manifest 分别把它们登记为公式密集和图片/视频案例。后续新样本直接放在本目录，不再继续追加到 `lala.md`。
+`cases/lala.md` 是第一批历史样本，目前包含两个非空 HTML 块。manifest 分别把它们登记为公式密集和图片/视频案例。它作为兼容种子保留原名；后续新内容使用独立文件，不再继续追加到 `lala.md`。
+
+`inbox/pig.md` 是尚未固化期望值的长图文样本，用来验证“先投递、后登记”的工作流。
