+++ features/rich-content/index.ts
@@ -1,0 +1,9 @@
+export {
+  LinkCard,
+  ZhihuContent,
+  type ZhihuContentProps,
+} from './components/ZhihuContent';
+export type {
+  TextSelectionInfo,
+  ZhihuDOMContentProps,
+} from './components/ZhihuDOMContent';
