+++ features/rich-content/tools/analyze-fixtures.mjs
@@ -1,28 +1,42 @@
 import { fileURLToPath } from 'node:url';
-import { analyzeFixtureCase, loadManifest } from './fixture-lib.mjs';
+import {
+  analyzeFixtureCase,
+  analyzeFixtureDirectory,
+  loadManifest,
+} from './fixture-lib.mjs';
 
 const manifestPath = fileURLToPath(
   new URL('../fixtures/manifest.json', import.meta.url),
 );
-const manifest = await loadManifest(manifestPath);
+const inboxPath = fileURLToPath(new URL('../fixtures/inbox/', import.meta.url));
+const analyzeInbox = process.argv.includes('--inbox');
 const requestedIds = new Set(
-  process.argv.slice(2).filter((argument) => argument !== '--json'),
+  process.argv
+    .slice(2)
+    .filter((argument) => !['--inbox', '--json'].includes(argument)),
 );
-const selectedCases = requestedIds.size
-  ? manifest.cases.filter((fixtureCase) => requestedIds.has(fixtureCase.id))
-  : manifest.cases;
+let results;
 
-if (selectedCases.length === 0) {
-  throw new Error(
-    `No matching fixtures: ${Array.from(requestedIds).join(', ')}`,
+if (analyzeInbox) {
+  results = await analyzeFixtureDirectory(inboxPath);
+} else {
+  const manifest = await loadManifest(manifestPath);
+  const selectedCases = requestedIds.size
+    ? manifest.cases.filter((fixtureCase) => requestedIds.has(fixtureCase.id))
+    : manifest.cases;
+
+  if (selectedCases.length === 0) {
+    throw new Error(
+      `No matching fixtures: ${Array.from(requestedIds).join(', ')}`,
+    );
+  }
+
+  results = await Promise.all(
+    selectedCases.map((fixtureCase) =>
+      analyzeFixtureCase(fixtureCase, manifestPath),
+    ),
   );
 }
-
-const results = await Promise.all(
-  selectedCases.map((fixtureCase) =>
-    analyzeFixtureCase(fixtureCase, manifestPath),
-  ),
-);
 
 if (process.argv.includes('--json')) {
   console.log(
@@ -47,7 +61,11 @@ if (process.argv.includes('--json')) {
       images: `${stats.activeImages}/${stats.totalImages}`,
       formulas: stats.formulaImages,
       videos: stats.videoBoxes,
-      validation: errors.length === 0 ? 'ok' : `${errors.length} error(s)`,
+      validation: analyzeInbox
+        ? 'unregistered'
+        : errors.length === 0
+          ? 'ok'
+          : `${errors.length} error(s)`,
     })),
   );
 }
