+++ package.json
@@ -9,6 +9,7 @@
     "ios": "expo run:ios",
     "web": "expo start --web",
     "analyze:rich-content": "node features/rich-content/tools/analyze-fixtures.mjs",
+    "analyze:rich-content:inbox": "node features/rich-content/tools/analyze-fixtures.mjs --inbox",
     "test:rich-content": "node --test features/rich-content/tests/*.test.mjs",
     "postinstall": "patch-package",
     "format": "biome format . --write",
